RUMM-2085 Migrate android bridge test files

+++ packages/core/android/src/main/java/com/datadog/reactnative/DdRum.kt
@@ -34,7 +34,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the view started (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun startView(key: String, name: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun startView(key: String, name: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
@@ -53,7 +53,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the view stopped (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun stopView(key: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun stopView(key: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
@@ -72,7 +72,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the action started (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun startAction(type: String, name: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun startAction(type: String, name: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
@@ -92,7 +92,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the action stopped (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun stopAction(type: String, name: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun stopAction(type: String, name: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
@@ -112,7 +112,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the action occurred (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun addAction(type: String, name: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun addAction(type: String, name: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
@@ -133,7 +133,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the resource started (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun startResource(key: String, method: String, url: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun startResource(key: String, method: String, url: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
@@ -156,7 +156,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the resource stopped (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun stopResource(key: String, statusCode: Double, kind: String, size: Double, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun stopResource(key: String, statusCode: Long, kind: String, size: Long, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
@@ -184,7 +184,7 @@ class DdRum(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(
      * @param timestampMs The timestamp when the error occurred (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun addError(message: String, source: String, stacktrace: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun addError(message: String, source: String, stacktrace: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val attributes = context.toHashMap().toMutableMap().apply {
             put(RumAttributes.INTERNAL_TIMESTAMP, timestampMs.toLong())
         }
