+++ packages/core/android/src/main/java/com/datadog/reactnative/DdTrace.kt
@@ -38,7 +38,7 @@ class DdTrace(reactContext: ReactApplicationContext, private val tracerProvider:
      * @param timestampMs The timestamp when the operation started (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun startSpan(operation: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun startSpan(operation: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val span = tracer.buildSpan(operation)
             .withStartTimestamp(TimeUnit.MILLISECONDS.toMicros(timestampMs.toLong()))
             .start()
@@ -59,7 +59,7 @@ class DdTrace(reactContext: ReactApplicationContext, private val tracerProvider:
      * @param timestampMs The timestamp when the operation stopped (in milliseconds). If not provided, current timestamp will be used.
      */
     @ReactMethod
-    fun finishSpan(spanId: String, context: ReadableMap, timestampMs: Double, promise: Promise) {
+    fun finishSpan(spanId: String, context: ReadableMap, timestampMs: Long, promise: Promise) {
         val span = spanMap.remove(spanId)
         if (span == null) {
             promise.resolve(null)
